:construction: refactor some stuff and add dictization
