+++ ckanext/dalrrd_emc_dcpr/helpers.py
@@ -11,7 +11,6 @@ from ckan.lib.helpers import build_nav_main as core_build_nav_main
 from . import constants
 from .logic.action.emc import show_version
 
-from .model.dcpr_request import DCPRRequestStatus
 
 logger = logging.getLogger(__name__)
 
@@ -232,20 +231,20 @@ def _pad_geospatial_extent(extent: typing.Dict, padding: float) -> typing.Dict:
 def get_status_labels() -> typing.Dict:
     """Get status labels for the DCPR requests"""
     status_labels = {
-        DCPRRequestStatus.UNDER_PREPARATION.value: (
+        constants.DCPRRequestStatus.UNDER_PREPARATION.value: (
             toolkit._("Under preparation"),
             "info",
         ),
-        DCPRRequestStatus.AWAITING_NSIF_REVIEW.value: (
+        constants.DCPRRequestStatus.AWAITING_NSIF_REVIEW.value: (
             toolkit._("Waiting for NSIF review"),
             "info",
         ),
-        DCPRRequestStatus.AWAITING_CSI_REVIEW.value: (
+        constants.DCPRRequestStatus.AWAITING_CSI_REVIEW.value: (
             toolkit._("Waiting for CSI review"),
             "info",
         ),
-        DCPRRequestStatus.ACCEPTED.value: (toolkit._("Accepted"), "success"),
-        DCPRRequestStatus.REJECTED.value: (toolkit._("Rejected"), "danger"),
+        constants.DCPRRequestStatus.ACCEPTED.value: (toolkit._("Accepted"), "success"),
+        constants.DCPRRequestStatus.REJECTED.value: (toolkit._("Rejected"), "danger"),
     }
 
     return status_labels
